+++ 11-test-react-mastery/public/index.html
@@ -2,19 +2,18 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="./poringbag.png" />
+    <link rel="icon" href="../src/images/poringbag.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="Web site created using create-react-app"
+      content="A Sample Youtube Web site created using create-react-app"
     />
 
     <link
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.4.1/semantic.min.css"
     />
-    <link rel="stylesheet" href="../src/components/App.css" />
 
     <title>Sample Youtube</title>
   </head>
