+++ 11-test-react-mastery/src/css/App.css
@@ -1,4 +1,4 @@
-body.custom {
+.custom {
   background-image: url("https://i.ibb.co/rcyn0rs/ro-background.jpg");
   object-fit: fill;
   background-size: 100% 100%;
